fixed number parsing, replaced hard coded typical value with correct dynamic one

+++ cppext/minimath.cpp
@@ -115,7 +115,7 @@ u64 truncateDecimals(u64 p19,unsigned digits){
 
 //uround and sround are coded to be like they will in optimized assembly
 u16 uround(float scaled){
-  if(scaled < 0.5) { //fp compares are the same cost as integer.
+  if(scaled < 0.5F) { //fp compares are the same cost as integer.
     return 0;
   }
   scaled *= 2; //expose rounding bit
@@ -127,14 +127,14 @@ u16 uround(float scaled){
 } /* uround */
 
 s16 sround(float scaled){ //#this would be so much cleaner and faster in asm!
-  if(scaled > 32766.5) {
+  if(scaled > 32766.5F) {
     return 32767;
   }
-  if(scaled < -32767.5) {
+  if(scaled < -32767.5F) {
     return -32768;
   }
   scaled += scaled >= 0 ? 0.5 : -0.5; //round away from 0. aka round the magnitude.
-  return int(scaled);
+  return s16(scaled);
 }
 
 int modulus(int value, unsigned cycle){
